refactor folders stracture
