+++ lib/core/routes/app_router.dart
@@ -5,12 +5,12 @@ import 'package:looqma/core/di/dependecy_injection.dart';
 import 'package:looqma/core/routes/routes.dart';
 import 'package:looqma/features/home/home_screen.dart';
 import 'package:looqma/features/login/data/repos/login_repo.dart';
-import 'package:looqma/features/login/views/cubit/login_cubit.dart';
-import 'package:looqma/features/login/views/presentation/login_screen.dart';
+import 'package:looqma/features/login/presentation/cubit/login_cubit.dart';
+import 'package:looqma/features/login/presentation/views/login_screen.dart';
 import 'package:looqma/features/otp_verify/data/repos/verfication_repo.dart';
-import 'package:looqma/features/otp_verify/views/cubit/verification_cubit.dart';
-import 'package:looqma/features/otp_verify/views/presentation/otp_verify_screen.dart';
-import 'package:looqma/features/sigh_up/views/presentation/sighn_up_screen.dart';
+import 'package:looqma/features/otp_verify/presentation/cubit/verification_cubit.dart';
+import 'package:looqma/features/otp_verify/presentation/views/otp_verify_screen.dart';
+import 'package:looqma/features/sigh_up/presentation/views/sighn_up_screen.dart';
 
 class AppRouter {
   static Route<void> getRoute(RouteSettings settings) {
