+++ lib/features/login/presentation/views/login_screen.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:looqma/core/utils/app_styles.dart';
-import 'package:looqma/features/login/views/presentation/widgets/donot_have_account.dart';
-import 'package:looqma/features/login/views/presentation/widgets/forget_password_button.dart';
-import 'package:looqma/features/login/views/presentation/widgets/login_form.dart';
-import 'package:looqma/features/login/views/presentation/widgets/sighn_in_button.dart';
+import 'package:looqma/features/login/presentation/views/widgets/donot_have_account.dart';
+import 'package:looqma/features/login/presentation/views/widgets/forget_password_button.dart';
+import 'package:looqma/features/login/presentation/views/widgets/login_form.dart';
+import 'package:looqma/features/login/presentation/views/widgets/sighn_in_button.dart';
 
 class LoginScreen extends StatelessWidget {
   const LoginScreen({super.key});
