+++ lib/features/login/presentation/views/widgets/email_text_form.dart
@@ -4,7 +4,7 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:looqma/core/common/widgets/custom_text_field.dart';
 import 'package:looqma/core/utils/app_styles.dart';
 import 'package:looqma/core/utils/my_validator.dart';
-import 'package:looqma/features/login/views/cubit/login_cubit.dart';
+import 'package:looqma/features/login/presentation/cubit/login_cubit.dart';
 
 class EmailTextForm extends StatelessWidget {
   const EmailTextForm({super.key});
