+++ lib/features/login/presentation/views/widgets/password_text_form.dart
@@ -5,7 +5,7 @@ import 'package:looqma/core/common/widgets/custom_text_field.dart';
 import 'package:looqma/core/utils/app_colors.dart';
 import 'package:looqma/core/utils/app_styles.dart';
 import 'package:looqma/core/utils/my_validator.dart';
-import 'package:looqma/features/login/views/cubit/login_cubit.dart';
+import 'package:looqma/features/login/presentation/cubit/login_cubit.dart';
 
 class PasswordTextForm extends StatefulWidget {
   const PasswordTextForm({super.key});
