+++ lib/features/otp_verify/presentation/views/otp_verify_screen.dart
@@ -3,8 +3,8 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:looqma/core/common/widgets/custom_back_arrow_app_bar.dart';
 import 'package:looqma/core/utils/app_assets.dart';
 import 'package:looqma/core/utils/app_styles.dart';
-import 'package:looqma/features/otp_verify/views/presentation/widgets/confirm_code_button.dart';
-import 'package:looqma/features/otp_verify/views/presentation/widgets/verify_code_widget.dart';
+import 'package:looqma/features/otp_verify/presentation/views/widgets/confirm_code_button.dart';
+import 'package:looqma/features/otp_verify/presentation/views/widgets/verify_code_widget.dart';
 import 'package:lottie/lottie.dart';
 
 class OtpVerifyScreen extends StatelessWidget {
